arm64: dts: qcom: add sdm670 and pixel 3a device trees
The Qualcomm Snapdragon 670 has been out for a while. Add a device tree
for it and the Google Pixel 3a as the first device.

The Pixel 3a has the same bootloader issue as the Pixel 3 and will not work
on Android 10 bootloaders or later until it gets fixed for the Pixel 3.

SoC Initial Features:
 - power management
 - clocks
 - pinctrl
 - eMMC
 - USB 2.0
 - GENI I2C ports 8-15
 - IOMMU
 - RPMh
 - interrupts

Device-Specific Initial Features:
 - side buttons (keys)
 - regulators
 - touchscreen

Signed-off-by: Richard Acayan <mailingradian@gmail.com>