drm/panel: add samsung s6e3fa7 panel driver
The S6E3FA7 is enabled in every Pixel 3a (non-XL) variant. Add the
driver for it, generated by linux-mdss-dsi-panel-driver-generator.

A commit on Google's msm kernel states that brightness values are
transferred in big endian, but the MIPI driver transmits values in
little endian. Before deciding on whether this should be changed,
perform a byte swap of the brightness before it enters the possibly
erroneous code.

Link: https://github.com/msm8916-mainline/linux-mdss-dsi-panel-driver-generator
Link: https://android.googlesource.com/kernel/msm/+/7fda1cd7b64710dafac5f34899611c6d35eb4cd2/arch/arm64/boot/dts/google/dsi-panel-s6e3fa7-1080p-cmd.dtsi
Link: https://android.googlesource.com/kernel/msm/+/754affd62d0ee268c686c53169b1dbb7deac8550
Signed-off-by: Richard Acayan <mailingradian@gmail.com>