arm64: dts: qcom: sdm670: add and enable adsp
The Audio DSP (Digital Signal Processor) is responsible for audio and
sensors on SDM670. Add this, along with the SMP2P nodes used to
communicate with it.

The sleep state is included to prevent the following crash:

    qcom_q6v5_pas remoteproc-adsp: fatal error received: err_qdi.c:450:EF:sensor_process:1:sar.cc:27:chre_utils fatal

Otherwise, it is unused.

No-submit: separate sargo patch, put smp2p nodes in another patch
Signed-off-by: Richard Acayan <mailingradian@gmail.com>