ASoC: qdsp6: q6voice-dai: reserve bit in shift attribute for rx/tx
The SDM670 and SDM845 kernel forks need to support the VoiceMMode1 path
type/voice mixer and use the shift register to support both VoiceMMode1
and CS-Voice. This creates a conflict when using the MSM8916 version of
the mixer controls patch, as the SDM670 kernel does.

Originally in SDM670, the logic was duplicated for both RX and TX, and
different functions have been referenced. This is not a great solution,
since the logic is almost the same except for a boolean passed to the
other functions.

Another idea was to indicate the path as RX/TX in a different register,
such as autodisable, invert, rshift, or sign_bit. This may have side
effects on the ASoC subsystem, which uses most of these fields for a
special purpose.

Reserving bit 0 in the shift attribute would have allowed the MSM8916
patch to be applied without additional fixups, but would also make it
difficult to add path types to the shift attribute, at best requiring
the use of macros for constructing the shift value.

Reserve bit 15 in the shift attribute to indicate the path as RX/TX.
This would reserve path types 32768-65535 when support for additional
path types is introduced.

Amends: ASoC: qdsp6: q6voice-dai: separate rx and tx mixer functions
Amends: ASoC: qdsp6: q6voice-dai: Introduce VoiceMMode1 dai
Amends: ASoC: qdsp6: q6voice-dai: get path type from shift param
Amends: ASoC: qdsp6: q6voice-dai: add internal mi2s
Signed-off-by: Richard Acayan <mailingradian@gmail.com>