arm64: configs: sdm670.config: disable some drivers and subsystems
Many options in the SDM670 and SDM845 kernel configuration fragments are
left unchanged because it gets difficult to maintain a comprehensive
list of unused drivers. This creates some bloat in the built kernel and
adds to compilation time. Disable these drivers and subsystems which
will likely never be used by end users on already-supported SDM670
devices.

The added list is a best-effort attempt and may erroneously include or
exclude entries. Drivers for devices on USB or Bluetooth are kept
because the end user may want to connect them. Some entries match
corresponding =y or =m entries in the SDM845 configuration because
SDM845 devices are not supported here.

Amends: arm64: configs: add sdm670 config fragment
Signed-off-by: Richard Acayan <mailingradian@gmail.com>