media: i2c: imx355: Support device tree probing and resource management
A device tree compatible makes it possible for this driver to be used on
Open Firmware devices. Initialization of resources such as the reset
GPIO and voltage regulators can be specified in the device tree and
handled by the driver. Add support for this so the Pixel 3a can use the
driver.

No-submit: split this into logical changes
Signed-off-by: Richard Acayan <mailingradian@gmail.com>