[WIP RFT] ASoC: qcom: make microphone work for real this time
The codec was emitting correct data, and the driver would almost work
perfectly downstream, but the Audio Front End was misconfigured. Even
though the clock rate was correct and frequencies of sounds from the
environment could be accurately measured, the samples were corrupted
because of the data delay value was misconfigured (set to 0 by kzalloc).
Configure the data delay as provided by the device tree and revert the
meaningless hacks.

Microphone enable sequence:

	cset "name='Sto1 ADC MIXL DMIC Switch' 1"
	cset "name='Sto2 ADC MIXL DMIC Switch' 1"

	cset "name='Stereo1 DMIC Mux' DMIC1"
	cset "name='Stereo2 DMIC Mux' DMIC1"

Signed-off-by: Richard Acayan <mailingradian@gmail.com>