media: camss: define dedicated resources without src clocks for sdm670
If a review comment from the mailing lists is acted on, the driver will
fail to probe unless the "_src" clocks are removed from the resources.
Define resources specifically for the SDM670 to use the ARRAY_SIZE()
macro for CSIPHY nodes and to remove these clocks.

The vfe0_src, vfe1_src, and vfe_lite_src clocks don't work at 19.2 MHz,
so remove that frequency from their children. The cphy_rx_src clock
needs to be set at 384 MHz, so set its children to that frequency.

Link: https://lore.kernel.org/linux-arm-msm/ZrP_4NRIfo71wDwo@radian
Amends: media: camss: add support for SDM670 camss
Signed-off-by: Richard Acayan <mailingradian@gmail.com>