media: i2c: imx363: Remove 1920_1080 mode
It currently results in a brownish tint for modes <= 1920x1080 and
breaks the mode in question - 1920x1080 - entirely in libcamera[1].

Using the 4k mode is not optimal either and results in a "zoom" effect
when apps request lower modes. Overall it arguably work betters though
and the image result is less confusing to downstream.

1: 1920x1080 apparently is not a valid mode for SRGGB10 and likely the
right crop size hase to be a bit bigger - to be investigated.

Signed-off-by: Robert Mader <robert.mader@collabora.com>
Signed-off-by: Richard Acayan <mailingradian@gmail.com>