arm64: dts: qcom: sdm670: remove camss endpoint nodes
There is no need to add these by default for all of SDM670. Originally,
they were added so there could be a label for each port. This is
unnecessary if the endpoints are all added in a fixup to the camss node.

Suggested-by: Bryan O'Donoghue <bryan.odonoghue@linaro.org>
Link: https://lore.kernel.org/r/488281f6-5e5d-4864-8220-63e2a0b2d7f2@linaro.org
Parent: arm64: dts: sdm670-google-common: add IMX363 rear camera
Signed-off-by: Richard Acayan <mailingradian@gmail.com>